fix: Restore forge_conformance goldens after typed AST refactor
- Add literal float promotion to cpp/python/go emit_*; Python single-quoted strings; matches the Rust/Kotlin convention so int literals in float context render with .0 across all five targets
- Refresh four stale goldens (Kotlin coercion direction in three validators/procedures, inline_mixed transform float literal) and add UPDATE_GOLDEN env var to assert_standalone_forge_lang/assert_inline_kinds for safe regeneration

+++ tests/forge/expected/CrossfileValidatorTransform.kt
@@ -12,7 +12,7 @@ class CrossfileValidatorTransform {
     fun validate(rawTemp: UShort): ValidationResult {
         if (rawTemp.toInt() < 0 || rawTemp.toInt() > 4095)
             return ValidationResult(false, "rawTemp_out_of_range")
-        if (!(computeTemperature(rawTemp.toInt()) > -40 && computeTemperature(rawTemp.toInt()) < 200))
+        if (!(computeTemperature(rawTemp) > -40 && computeTemperature(rawTemp) < 200))
             return ValidationResult(false, "plausibility_failed")
         return ValidationResult(true, "")
     }
